Do not prune the move Kg7xNh8 in qsearch
This is an illustrative fix for the following drawn position:

./stockfish
position fen 8/6k1/8/8/8/3B3P/6KP/8 w - - 0 1
go depth 30

The endgame KBNPK is not scored as a KNOWN_WIN so if white underpromotes
to knight the move Kg7xNh8 can be pruned.

White can delay the underpromotion to knight almost indefinitely so the
main search never overcomes the error in qsearch.

Here are two possible fixes:

1. Disable futility pruning in qsearch with only pawns for side to move.

2. Award VALUE_KNOWN_WIN in the KXK endgame if we have B, N and P.