Pawn imbalance change
Discourage pawn trades when we are ahead in pawns.
Encourage pawn trades when we are behind in pawns.

LinearCoeff[PAWN] is reduced by 8 times more
than QuadCoeff[PAWN][PAWN] is raised.

So imbalance changes by:

-8 * delta * (w - b) + delta * (w^2 -b^2)

which is precisely

delta * (w - b) * (w + b - 8)

ie. no change when the total number of pawns is 8.