In Material::probe() search for scalingFunction last
If we find a specialised evaluation function in Material::probe() then
we return immediately. This is justified because evaluation will not
use any of the other values in Material::Entry.

However, if we only find a scalingFunction then imbalance (and possibly
factor) will be used even though they refer to the old entry. Search for
a scalingFunction last to prevent this behaviour.

Bench: 5321504